H.Prelude

Documentation

module Control.Applicative

module Control.Category

module Control.Monad

module Control.Monad.Except

module Control.Monad.Identity

module Control.Monad.Reader

module Control.Monad.State

module Data.Bool

module Data.ByteString

module Data.Char

module Data.Either

module Data.Foldable

module Data.Function

module Data.List

module Data.Map

module Data.Maybe

module Data.Monoid

module Data.Ord

module Data.Ratio

module Data.Set

module Data.Text

module Data.Text.Encoding

module Data.Traversable

module Data.Tuple

module Debug.Trace

module Prelude

show :: Show a => a -> Text

read :: Read a => Text -> Maybe a

todo :: a

error :: Text -> a

impossible :: Text -> a

whenJust :: Monad m => Maybe a -> (a -> m ()) -> m ()

If the input is Just, do a monadic action on the value

onFst :: (a -> c) -> (a, b) -> (c, b)

onSnd :: (b -> c) -> (a, b) -> (a, c)

onFstF :: Functor f => (a -> f c) -> (a, b) -> f (c, b)

onSndF :: Functor f => (b -> f c) -> (a, b) -> f (a, c)

lift2 :: (MonadTrans t, MonadTrans u, Monad m, Monad (u m)) => m a -> t (u m) a

lift3 :: (MonadTrans t, MonadTrans u, MonadTrans v, Monad m, Monad (u m), Monad (t (u m))) => m a -> v (t (u m)) a

lift4 :: (MonadTrans t, MonadTrans u, MonadTrans v, MonadTrans w, Monad m, Monad (u m), Monad (t (u m)), Monad (v (t (u m)))) => m a -> w (v (t (u m))) a

modifyM :: Monad m => (s -> m s) -> StateT s m ()

Modify the state of a StateT using a monadic action of the inner monad.

minimumByM :: Monad m => (a -> a -> m Ordering) -> [a] -> m a

Find the minimum element of a list using a monadic comparison action.

data Proxy a

Constructors

Proxy

asProxied :: a -> Proxy a -> a infix 8

eitherAlt :: Alternative f => f a -> f b -> f (Either a b) infixl 3

Like |, but the operands may have different value types, with Either providing a union of those two types in the result

sequenceWhileJust :: Monad m => [m (Maybe a)] -> m [a]

Sequence a list of actions that return Maybes, stopping at the first Nothing

headView :: [a] -> Maybe (a, [a])

Deconstruct a list into its head and tail

fst3 :: (a, b, c) -> a

Get the first element of a triple

snd3 :: (a, b, c) -> b

Get the second element of a triple

thd3 :: (a, b, c) -> c

Get the third element of a triple

unionWithM :: (Ord k, Monad m) => (a -> a -> m a) -> Map k a -> Map k a -> m (Map k a)

Union two maps, with a monadic action for merging duplicates

whenM :: Monad m => m Bool -> m () -> m ()

Like when, but the condition is also a monadic action

traceVal :: Show a => a -> a

setCatMaybes :: Ord a => Set (Maybe a) -> Set a

setSequence :: (Ord a, Applicative f) => Set (f a) -> f (Set a)

onLeft :: (a -> c) -> Either a b -> Either c b

onRight :: (b -> c) -> Either a b -> Either a c

whenJustM :: Monad m => m (Maybe a) -> (a -> m ()) -> m ()

boolToMaybe :: Bool -> Maybe ()