H.Pretty

Documentation

fsep :: [Doc] -> Doc

"Paragraph fill" version of sep.

fcat :: [Doc] -> Doc

"Paragraph fill" version of cat.

cat :: [Doc] -> Doc

Either hcat or vcat.

sep :: [Doc] -> Doc

Either hsep or vcat.

(<+>) :: Doc -> Doc -> Doc infixl 6

Beside, separated by space, unless one of the arguments is empty. <+> is associative, with identity empty.

(<>) :: Doc -> Doc -> Doc infixl 6

Beside. <> is associative, with identity empty.

($+$) :: Doc -> Doc -> Doc infixl 5

Above, with no overlapping. $+$ is associative, with identity empty.

($$) :: Doc -> Doc -> Doc infixl 5

Above, except that if the last line of the first argument stops at least one position before the first line of the second begins, these two lines are overlapped. For example:

   text "hi" $$ nest 5 (text "there")

lays out as

   hi   there

rather than

   hi
        there

$$ is associative, with identity empty, and also satisfies

(x $$ y) <> z = x $$ (y <> z), if y non-empty.

punctuate :: Doc -> [Doc] -> [Doc]

punctuate p [d1, ... dn] = [d1 <> p, d2 <> p, ... dn-1 <> p, dn]

hang :: Doc -> Int -> Doc -> Doc

hang d1 n d2 = sep [d1, nest n d2]

nest :: Int -> Doc -> Doc

Nest (or indent) a document by a given number of positions (which may also be negative). nest satisfies the laws:

nest 0 x = x

nest k (nest k' x) = nest (k+k') x

nest k (x <> y) = nest k x <> nest k y

nest k (x $$ y) = nest k x $$ nest k y

nest k empty = empty

x <> nest k y = x <> y, if x non-empty

The side condition on the last law is needed because empty is a left identity for <>.

vcat :: [Doc] -> Doc

List version of $$.

hsep :: [Doc] -> Doc

List version of <+>.

hcat :: [Doc] -> Doc

List version of <>.

braces

Arguments

:: Doc

-> Doc
Wrap document in {...}

brackets

Arguments

:: Doc

-> Doc
Wrap document in [...]

parens

Arguments

:: Doc

-> Doc
Wrap document in (...)

doubleQuotes

Arguments

:: Doc

-> Doc
Wrap document in "..."

quotes

Arguments

:: Doc

-> Doc
Wrap document in '...'

rational

Arguments

:: Rational

-> Doc
rational n = text (show n)

double

Arguments

:: Double

-> Doc
double n = text (show n)

float

Arguments

:: Float

-> Doc
float n = text (show n)

integer

Arguments

:: Integer

-> Doc
integer n = text (show n)

int

Arguments

:: Int

-> Doc
int n = text (show n)

rbrace

Arguments

:: Doc
A '}' character

lbrace

Arguments

:: Doc
A '{' character

rbrack

Arguments

:: Doc
A ']' character

lbrack

Arguments

:: Doc
A '[' character

rparen

Arguments

:: Doc
A ')' character

lparen

Arguments

:: Doc
A '(' character

equals

Arguments

:: Doc
A '=' character

space

Arguments

:: Doc
A space character

colon

Arguments

:: Doc
A : character

comma

Arguments

:: Doc
A ',' character

semi

Arguments

:: Doc
A ';' character

isEmpty :: Doc -> Bool

Returns True if the document is empty

empty :: Doc

The empty document, with no height and no width. empty is the identity for <>, <+>, $$ and $+$, and anywhere in the argument list for sep, hcat, hsep, vcat, fcat etc.

char :: Char -> Doc

A document of height and width 1, containing a literal character.

data Doc

The abstract type of documents. A Doc represents a set of layouts. A Doc with no occurrences of Union or NoDoc represents just one layout.

style :: Style

The default style (mode=PageMode, lineLength=100, ribbonsPerLine=1.5).

data Style

A rendering style. Allows us to specify constraints to choose among the many different rendering options.

Constructors

Style

Fields

mode :: Mode

The rendering mode.

lineLength :: Int

Maximum length of a line, in characters.

ribbonsPerLine :: Float

Ratio of line length to ribbon length. A ribbon refers to the characters on a line excluding indentation. So a lineLength of 100, with a ribbonsPerLine of 2.0 would only allow up to 50 characters of ribbon to be displayed on a line, while allowing it to be indented up to 50 characters.

data Mode

Rendering mode.

Constructors

PageMode
Normal rendering (lineLength and ribbonsPerLine respected').

ZigZagMode
With zig-zag cuts.

LeftMode
No indentation, infinitely long lines (lineLength ignored), but explicit new lines, i.e., text "one" $$ text "two", are respected.

OneLineMode
All on one line, lineLength ignored and explicit new lines ($$) are turned into spaces.

data TextDetails

Constructors

Chr !Char

Str Text

text :: Text -> Doc

sizedText :: Int -> Text -> Doc

zeroWidthText :: Text -> Doc

render :: Doc -> Text

renderStyle :: Style -> Doc -> Text

fullRender :: Mode -> Int -> Float -> (TextDetails -> a -> a) -> a -> Doc -> a